slice(blorktools): handle drag move
